+++ abstractions/php/src/Serialization/ParseNode.php
@@ -5,7 +5,6 @@ namespace Microsoft\Kiota\Abstractions\Serialization;
 use DateInterval;
 use DateTime;
 use Microsoft\Kiota\Abstractions\Enum;
-use Microsoft\Kiota\Abstractions\Types\Byte;
 use Microsoft\Kiota\Abstractions\Types\Date;
 use Microsoft\Kiota\Abstractions\Types\Time;
 use Psr\Http\Message\StreamInterface;
@@ -44,16 +43,16 @@ interface ParseNode {
 
     /**
      * Gets the model object value of the node.
-     * @param string $type The type for the Parsable object.
+     * @param array{string,string} $type The type for the Parsable object.
      * @return Parsable|null the model object value of the node.
      */
-    public function getObjectValue(string $type): ?Parsable;
+    public function getObjectValue(array $type): ?Parsable;
 
     /**
-     * @param string $type The underlying type for the Parsable class.
+     * @param array{string,string} $type The underlying type for the Parsable class.
      * @return array<Parsable>|null An array of Parsable values.
      */
-    public function getCollectionOfObjectValues(string $type): ?array;
+    public function getCollectionOfObjectValues(array $type): ?array;
 
     /**
      * Get a collection of values that are not parsable in Nature.
@@ -92,12 +91,6 @@ interface ParseNode {
      * @return Enum|null the Enum value of the node.
      */
     public function getEnumValue(string $targetEnum): ?Enum;
-
-    /**
-     * Return a byte value.
-     * @return Byte|null
-     */
-    public function getByteValue(): ?Byte;
 
     /**
      * Get a Stream from node.